Consolidate renderers
